Update TrappedMacros.py
Previous commits were not commited properly, and the commit history is lost. various new functionalities added. Autofarm, and other functions added to support main functionalities of the program. Findbobber overhauled to find any object and return its cords. new process to detect if esc is pressed anytime during the program to return to the gui and exit the current processes. Various bug fixes and improvements not mentioned due to the commit history being lost